refactor & reorganize for modularity
