+++ data/src/chart.rs
@@ -8,10 +8,6 @@ pub mod timeandsales;
 
 pub use kline::KlineChartKind;
 
-pub fn round_to_tick(value: f32, tick_size: f32) -> f32 {
-    (value / tick_size).round() * tick_size
-}
-
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct ChartLayout {
     pub crosshair: bool,
