+++ data/src/lib.rs
@@ -4,6 +4,7 @@ pub mod chart;
 pub mod config;
 pub mod layout;
 pub mod log;
+pub mod util;
 
 use std::fs::File;
 use std::io::{Read, Write};
