+++ src/chart/indicator.rs
@@ -6,8 +6,9 @@ use iced::{
     widget::canvas::{self, Cache, Geometry},
 };
 
-use super::{abbr_large_numbers, round_to_tick, scale::linear};
+use super::scale::linear;
 use crate::chart::scale::{AxisLabel, Label, calc_label_rect};
+use data::util::{abbr_large_numbers, round_to_tick};
 
 use super::{Interaction, Message};
