+++ src/chart/indicator/open_interest.rs
@@ -4,10 +4,9 @@ use iced::widget::canvas::{self, Cache, Event, Geometry, Path, Stroke};
 use iced::widget::{Canvas, center, container, row, text, vertical_rule};
 use iced::{Element, Length, Point, Rectangle, Renderer, Size, Theme, Vector, mouse};
 
-use crate::chart::{
-    Basis, Caches, CommonChartData, Interaction, Message, format_with_commas, round_to_tick,
-};
+use crate::chart::{Basis, Caches, CommonChartData, Interaction, Message};
 use crate::style::{self, get_dashed_line};
+use data::util::{format_with_commas, round_to_tick};
 use exchange::Timeframe;
 
 pub fn create_indicator_elem<'a>(
