+++ src/screen/dashboard.rs
@@ -1,4 +1,5 @@
 pub mod pane;
+pub mod panel;
 pub mod theme_editor;
 pub mod tickers_table;
